[TICA|N-TICA] derived from new base class to avoid copy-paste code.
